auction_project/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
bootstrap4==0.1.0
contourpy==1.3.3
crispy==0.8.0
crispy-bootstrap5==2025.6
cycler==0.12.1
Django==5.2.5
django-crispy-forms==2.5
django-ninja==1.4.3
django-widget-tweaks==1.5.0
fabio==2025.10.0
factory_boy==3.3.3
Faker==38.2.0
fonttools==4.61.0
h5py==3.15.1
hdf5plugin==6.0.0
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==11.3.0
platformdirs==4.5.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.5
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
python-dateutil==2.9.0.post0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scipy==1.16.3
shell==1.0.1
silx==2.2.2
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
xraydb==4.5.8
celery>=5.5,<6.0
redis>=4.6.0

